# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Assignment Workspace for Coding Test DELOS Aqua.",
        "title": "AquaFarm Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Fadli Zul Fahmi",
            "url": "http://www.linkedin.com/in/fadli-zul-fahmi",
            "email": "hi@fadli.dev"
        },
        "version": "0.6"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/v1/logs": {
            "get": {
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "description": "Get All Logs Data From Logs Middleware.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Logs"
                ],
                "summary": "Get All Logs Data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPResponseDataLogs"
                        },
                        "headers": {
                            "Authorization": {
                                "type": "string",
                                "description": "apiKey"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPResponseAction"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPResponseAction"
                        },
                        "headers": {
                            "Authorization": {
                                "type": "string",
                                "description": "apiKey"
                            }
                        }
                    }
                }
            }
        },
        "/v1/logs/statistics": {
            "get": {
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "description": "Get Statistics Data From Logs Middleware.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Logs"
                ],
                "summary": "Get Statistics Data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPResponseDataStatistics"
                        },
                        "headers": {
                            "Authorization": {
                                "type": "string",
                                "description": "apiKey"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPResponseAction"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HTTPResponseAction"
                        },
                        "headers": {
                            "Authorization": {
                                "type": "string",
                                "description": "apiKey"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Logs": {
            "type": "object",
            "properties": {
                "access_id": {
                    "type": "string"
                },
                "access_name": {
                    "type": "string"
                },
                "accessed_at": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "request": {
                    "type": "string"
                },
                "response": {
                    "type": "string"
                },
                "sequence": {
                    "type": "string"
                },
                "user_agent": {
                    "type": "string"
                }
            }
        },
...
        "model.StatisticsData": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "response_200": {
                    "type": "integer"
                },
                "response_404": {
                    "type": "integer"
                },
                "response_500": {
                    "type": "integer"
                },
                "response_etc": {
                    "type": "integer"
                },
                "unique_user_agent": {
                    "type": "integer"
                }
            }
        },
        "model.StatisticsGroup": {
            "type": "object",
            "properties": {
                "request": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.StatisticsData"
                    }
                }
            }
        },
        "response.HTTPResponseAction": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "response.HTTPResponseDataLogs": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Logs"
                    }
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "response.HTTPResponseDataStatistics": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.StatisticsGroup"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "APIKeyHeader": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}